refactor: basic 과제
