+++ src/basic/constants/index.ts
@@ -1,0 +1,51 @@
+export const PRODUCT_IDS = {
+  KEYBOARD: "p1",
+  MOUSE: "p2",
+  MONITOR_ARM: "p3",
+  LAPTOP_POUCH: "p4",
+  SPEAKER: "p5",
+} as const;
+
+export const DISCOUNT_RULES = {
+  ITEM_DISCOUNT_THRESHOLD: 10,
+  ITEM_DISCOUNT_RATES: {
+    [PRODUCT_IDS.KEYBOARD]: 10,
+    [PRODUCT_IDS.MOUSE]: 15,
+    [PRODUCT_IDS.MONITOR_ARM]: 20,
+    [PRODUCT_IDS.LAPTOP_POUCH]: 5,
+    [PRODUCT_IDS.SPEAKER]: 25,
+  },
+  BULK_DISCOUNT_THRESHOLD: 30,
+  BULK_DISCOUNT_RATE: 25,
+  SPECIAL_DISCOUNT_DAYS: [2],
+  SPECIAL_DISCOUNT_RATE: 10,
+  LIGHTNING_SALE_RATE: 20,
+  RECOMMENDATION_DISCOUNT_RATE: 5,
+};
+
+export const STOCK_THRESHOLDS = {
+  LOW_STOCK_WARNING: 5,
+  TOTAL_STOCK_WARNING: 50,
+  TOTAL_STOCK_CRITICAL: 30,
+} as const;
+
+export const POINTS_RULES = {
+  BASE_CALCULATION_UNIT: 1000,
+  SPECIAL_POINTS_DAYS: [2],
+  SPECIAL_POINTS_MULTIPLIER: 2,
+  COMBO_BONUS: {
+    KEYBOARD_MOUSE: 50,
+    FULL_SET: 100,
+  },
+  QUANTITY_BONUS: {
+    TIER_1: { threshold: 10, bonus: 20 },
+    TIER_2: { threshold: 20, bonus: 50 },
+    TIER_3: { threshold: 30, bonus: 100 },
+  },
+};
+
+export const SALE_INTERVALS = {
+  LIGHTNING_SALE_INTERVAL: 30000,
+  RECOMMENDATION_INTERVAL: 60000,
+  LIGHTNING_SALE_INITIAL_DELAY: 10000,
+} as const;
