+++ src/basic/main.basic.ts
@@ -1,1330 +1,17 @@
-/**
- * DOM 요소 및 상태 관리자 (React useRef/useState 패턴)
- * React 변환 시 useRef/useState로 직접 변환 가능
- */
-const useDOMManager = {
-  elements: {
-    stockStatus: null,
-    productSelect: null,
-    addToCartButton: null,
-    cartDisplay: null,
-  },
-
-  state: {
-    lastSelectedProductId: null,
-  },
-
-  getElement(elementName) {
-    return this.elements[elementName];
-  },
-
-  setElement(elementName, element) {
-    this.elements[elementName] = element;
-  },
-
-  getState(stateName) {
-    return this.state[stateName];
-  },
-
-  setState(stateName, value) {
-    this.state[stateName] = value;
-  },
-};
-
-const PRODUCT_IDS = {
-  KEYBOARD: "p1",
-  MOUSE: "p2",
-  MONITOR_ARM: "p3",
-  LAPTOP_POUCH: "p4",
-  SPEAKER: "p5",
-};
-
-const DISCOUNT_RULES = {
-  ITEM_DISCOUNT_THRESHOLD: 10,
-  ITEM_DISCOUNT_RATES: {
-    [PRODUCT_IDS.KEYBOARD]: 10,
-    [PRODUCT_IDS.MOUSE]: 15,
-    [PRODUCT_IDS.MONITOR_ARM]: 20,
-    [PRODUCT_IDS.LAPTOP_POUCH]: 5,
-    [PRODUCT_IDS.SPEAKER]: 25,
-  },
-
-  BULK_DISCOUNT_THRESHOLD: 30,
-  BULK_DISCOUNT_RATE: 25,
-
-  SPECIAL_DISCOUNT_DAYS: [2],
-  SPECIAL_DISCOUNT_RATE: 10,
-
-  LIGHTNING_SALE_RATE: 20,
-  RECOMMENDATION_DISCOUNT_RATE: 5,
-};
-
-const STOCK_THRESHOLDS = {
-  LOW_STOCK_WARNING: 5,
-  TOTAL_STOCK_WARNING: 50,
-  TOTAL_STOCK_CRITICAL: 30,
-};
-
-const POINTS_RULES = {
-  BASE_CALCULATION_UNIT: 1000,
-
-  SPECIAL_POINTS_DAYS: [2],
-  SPECIAL_POINTS_MULTIPLIER: 2,
-
-  COMBO_BONUS: {
-    KEYBOARD_MOUSE: 50,
-    FULL_SET: 100,
-  },
-
-  QUANTITY_BONUS: {
-    TIER_1: { threshold: 10, bonus: 20 },
-    TIER_2: { threshold: 20, bonus: 50 },
-    TIER_3: { threshold: 30, bonus: 100 },
-  },
-};
-
-/**
- * 특별 할인 요일 체크
- * @param {Date} date
- * @returns {boolean}
- */
-const isSpecialDiscountDay = (date: Date = new Date()): boolean => {
-  return DISCOUNT_RULES.SPECIAL_DISCOUNT_DAYS.includes(date.getDay());
-};
-
-/**
- * 특별 포인트 요일 체크
- * @param {Date} date
- * @returns {boolean}
- */
-const isSpecialPointsDay = (date: Date = new Date()): boolean => {
-  return POINTS_RULES.SPECIAL_POINTS_DAYS.includes(date.getDay());
-};
-
-/**
- * 요일 이름 추출
- * @param {number} dayIndex
- * @returns {string} 요일 이름
- */
-const getKoreanDayName = (dayIndex: number): string => {
-  if (dayIndex === 0) return "일요일";
-  if (dayIndex === 1) return "월요일";
-  if (dayIndex === 2) return "화요일";
-  if (dayIndex === 3) return "수요일";
-  if (dayIndex === 4) return "목요일";
-  if (dayIndex === 5) return "금요일";
-  if (dayIndex === 6) return "토요일";
-  return "";
-};
-
-const SALE_INTERVALS = {
-  LIGHTNING_SALE_INTERVAL: 30000,
-  RECOMMENDATION_INTERVAL: 60000,
-  LIGHTNING_SALE_INITIAL_DELAY: 10000,
-};
-
-interface IProduct {
-  id: string;
-  name: string;
-  val: number;
-  originalVal: number;
-  q: number;
-  onSale: boolean;
-  suggestSale: boolean;
-}
-
-interface ICartItem {
-  name: string;
-  quantity: number;
-  itemTotal: number;
-}
-
-interface IItemDiscount {
-  name: string;
-  discount: number;
-}
-
-interface IDiscountData {
-  hasDiscount: boolean;
-  savedAmount: number;
-  discountPercentage: string;
-  formattedSavedAmount: string;
-}
-
-interface ICartCalculation {
-  subtotal: number;
-  itemCount: number;
-  itemDiscounts: IItemDiscount[];
-  totalAmount: number;
-  discountRate: number;
-  originalTotal: number;
-  isSpecialDiscount: boolean;
-}
-
-interface IProductSaleStatus {
-  val?: number;
-  onSale?: boolean;
-  suggestSale?: boolean;
-}
-
-interface IBonusPointsResult {
-  totalPoints: number;
-  details: string[];
-  breakdown: {
-    base: number;
-    specialDay: any;
-    combo: any;
-    quantity: any;
-  };
-}
-
-interface ICartItemData {
-  itemIndex: number;
-  priceHTML: string;
-  nameText: string;
-  priceClassName: string;
-  isDiscounted: boolean;
-}
-
-const useProductData = {
-  products: [
-    {
-      id: PRODUCT_IDS.KEYBOARD,
-      name: "버그 없애는 키보드",
-      val: 10000,
-      originalVal: 10000,
-      q: 50,
-      onSale: false,
-      suggestSale: false,
-    },
-    {
-      id: PRODUCT_IDS.MOUSE,
-      name: "생산성 폭발 마우스",
-      val: 20000,
-      originalVal: 20000,
-      q: 30,
-      onSale: false,
-      suggestSale: false,
-    },
-    {
-      id: PRODUCT_IDS.MONITOR_ARM,
-      name: "거북목 탈출 모니터암",
-      val: 30000,
-      originalVal: 30000,
-      q: 20,
-      onSale: false,
-      suggestSale: false,
-    },
-    {
-      id: PRODUCT_IDS.LAPTOP_POUCH,
-      name: "에러 방지 노트북 파우치",
-      val: 15000,
-      originalVal: 15000,
-      q: 0,
-      onSale: false,
-      suggestSale: false,
-    },
-    {
-      id: PRODUCT_IDS.SPEAKER,
-      name: `코딩할 때 듣는 Lo-Fi 스피커`,
-      val: 25000,
-      originalVal: 25000,
-      q: 10,
-      onSale: false,
-      suggestSale: false,
-    },
-  ],
-
-  /**
-   * 상품 목록 반환
-   * @returns {Array} 상품 목록 배열
-   */
-  getProducts(): IProduct[] {
-    return [...this.products];
-  },
-
-  /**
-   * 총 재고 계산
-   * @returns {number} 총 재고 수량
-   */
-  getTotalStock(): number {
-    return this.products.reduce((total, product) => total + product.q);
-  },
-
-  /**
-   * 상품 ID로 상품 찾기
-   * @param {string} id - 상품 ID
-   * @returns {Object|null} 찾은 상품 객체 또는 null
-   */
-  findProductById(id: string): IProduct | null {
-    return this.products.find((product) => product.id === id) || null;
-  },
-
-  /**
-   * 상품 재고 수량 업데이트 (불변 업데이트)
-   * @param {string} id - 상품 ID
-   * @param {number} stockChange - 재고 변경량 (음수면 감소, 양수면 증가)
-   * @returns {boolean} 업데이트 성공 여부
-   */
-  updateProductStock(id: string, stockChange: number): boolean {
-    const productIndex = this.products.findIndex((product) => product.id === id);
-    if (productIndex === -1) {
-      return false;
-    }
-
-    const currentProduct = this.products[productIndex];
-    const newStock = currentProduct.q + stockChange;
-
-    if (newStock < 0) {
-      return false;
-    }
-
-    this.products[productIndex] = {
-      ...currentProduct,
-      q: newStock,
-    };
-
-    return true;
-  },
-
-  /**
-   * 상품 가격 업데이트 (불변 업데이트)
-   * @param {string} id - 상품 ID
-   * @param {number} newPrice - 새로운 가격
-   * @returns {boolean} 업데이트 성공 여부
-   */
-  updateProductPrice(id: string, newPrice: number): boolean {
-    const productIndex = this.products.findIndex((product) => product.id === id);
-    if (productIndex === -1) {
-      return false;
-    }
-
-    const currentProduct = this.products[productIndex];
-
-    this.products[productIndex] = {
-      ...currentProduct,
-      val: newPrice,
-    };
-
-    return true;
-  },
-
-  /**
-   * 상품 세일 상태 업데이트 (불변 업데이트)
-   * @param {string} id - 상품 ID
-   * @param {Object} saleUpdates - 세일 상태 업데이트 객체
-   * @param {boolean} [saleUpdates.onSale] - 번개세일 상태
-   * @param {boolean} [saleUpdates.suggestSale] - 추천세일 상태
-   * @returns {boolean} 업데이트 성공 여부
-   */
-  updateProductSaleStatus(id: string, saleUpdates: { onSale: boolean; suggestSale: boolean }): boolean {
-    const productIndex = this.products.findIndex((product) => product.id === id);
-    if (productIndex === -1) {
-      return false;
-    }
-
-    const currentProduct = this.products[productIndex];
-
-    this.products[productIndex] = {
-      ...currentProduct,
-      ...saleUpdates,
-    };
-
-    return true;
-  },
-
-  /**
-   * 상품 추천 할인 적용 (불변 업데이트)
-   * @param {string} id - 상품 ID
-   * @param {number} discountRate - 할인율 (백분율)
-   * @returns {boolean} 업데이트 성공 여부
-   */
-  applyRecommendationDiscount(id: string, discountRate: number): boolean {
-    const productIndex = this.products.findIndex((product) => product.id === id);
-    if (productIndex === -1) {
-      return false;
-    }
-
-    const currentProduct = this.products[productIndex];
-    const discountedPrice = Math.round((currentProduct.val * (100 - discountRate)) / 100);
-
-    this.products[productIndex] = {
-      ...currentProduct,
-      val: discountedPrice,
-      suggestSale: true,
-    };
-
-    return true;
-  },
-};
-
-const useStockManager = {
-  /**
-   * 재고 경고 메시지 생성
-   * @returns {string} 재고 경고 메시지
-   */
-  generateStockWarningMessage(): string {
-    const products = useProductData.getProducts();
-    let warningMsg = "";
-
-    products.forEach((item) => {
-      if (item.q < STOCK_THRESHOLDS.LOW_STOCK_WARNING) {
-        if (item.q > 0) {
-          warningMsg += `${item.name}: 재고 부족 (${item.q}개 남음)\n`;
-        } else {
-          warningMsg += `${item.name}: 품절\n`;
-        }
-      }
-    });
-
-    return warningMsg;
-  },
-
-  /**
-   * 재고 정보 UI 업데이트
-   */
-  updateStockInfoDisplay(): void {
-    const warningMessage = this.generateStockWarningMessage();
-    const stockInfoElement = document.getElementById("stock-status");
-    if (stockInfoElement) {
-      stockInfoElement.textContent = warningMessage;
-    }
-  },
-};
-
-const useCartManager = {
-  totalAmount: 0,
-  itemCount: 0,
-
-  /**
-   * 장바구니 총 금액 반환
-   * @returns {number} 총 금액
-   */
-  getTotalAmount() {
-    return this.totalAmount;
-  },
-
-  /**
-   * 장바구니 총 상품 개수 반환
-   * @returns {number} 상품 개수
-   */
-  getItemCount() {
-    return this.itemCount;
-  },
-
-  /**
-   * 장바구니 상태 초기화
-   */
-  resetCart() {
-    this.totalAmount = 0;
-    this.itemCount = 0;
-  },
-
-  /**
-   * 장바구니 총액과 개수 설정 (내부용)
-   * @param {number} amount - 총 금액
-   * @param {number} count - 상품 개수
-   */
-  setCartTotals(amount, count) {
-    this.totalAmount = amount;
-    this.itemCount = count;
-  },
-
-  /**
-   * 장바구니 아이템들로부터 총액과 개수 계산
-   * @param {HTMLCollection} cartItems - 장바구니 DOM 아이템들
-   * @returns {Object} 계산 결과 {subtotal, itemCount, itemDiscounts}
-   */
-  calculateCartTotals(cartItems: HTMLCollection) {
-    let subtotal = 0;
-    let itemCount = 0;
-    const itemDiscounts: IItemDiscount[] = [];
-
-    for (let i = 0; i < cartItems.length; i += 1) {
-      const cartItem = cartItems[i] as HTMLElement;
-      const curItem = useProductData.findProductById(cartItem.id);
-
-      if (curItem) {
-        const qtyElem = cartItem.querySelector(".quantity-number") as HTMLElement;
-        if (qtyElem?.textContent) {
-          const q = parseInt(qtyElem.textContent, 10);
-          const itemTot = curItem.val * q;
-
-          itemCount += q;
-          subtotal += itemTot;
-
-          if (q >= DISCOUNT_RULES.ITEM_DISCOUNT_THRESHOLD) {
-            const disc = (DISCOUNT_RULES.ITEM_DISCOUNT_RATES[curItem.id] || 0) / 100;
-            if (disc > 0) {
-              itemDiscounts.push({
-                name: curItem.name,
-                discount: DISCOUNT_RULES.ITEM_DISCOUNT_RATES[curItem.id],
-              });
-            }
-          }
-        }
-      }
-    }
-
-    return {
-      subtotal,
-      itemCount,
-      itemDiscounts,
-    };
-  },
-
-  /**
-   * 총 할인율 계산 (대량 구매 할인, 화요일 할인 포함)
-   * @param {number} subtotal - 소계
-   * @param {number} itemCount - 상품 개수
-   * @param {Array} itemDiscounts - 개별 상품 할인 목록
-   * @returns {Object} {totalAmount, discountRate, originalTotal}
-   */
-  calculateFinalAmount(subtotal, itemCount, itemDiscounts) {
-    let totalAmount = subtotal;
-    let discountRate = 0;
-    const originalTotal = subtotal;
-
-    if (itemCount < DISCOUNT_RULES.BULK_DISCOUNT_THRESHOLD) {
-      itemDiscounts.forEach((item) => {
-        const discountAmount = subtotal * (item.discount / 100);
-        totalAmount -= discountAmount;
-      });
-      discountRate = (subtotal - totalAmount) / subtotal;
-    }
-
-    if (itemCount >= DISCOUNT_RULES.BULK_DISCOUNT_THRESHOLD) {
-      totalAmount = subtotal * (1 - DISCOUNT_RULES.BULK_DISCOUNT_RATE / 100);
-      discountRate = DISCOUNT_RULES.BULK_DISCOUNT_RATE / 100;
-    }
-
-    const today = new Date();
-    const isSpecialDiscount = isSpecialDiscountDay(today);
-    if (isSpecialDiscount && totalAmount > 0) {
-      totalAmount *= 1 - DISCOUNT_RULES.SPECIAL_DISCOUNT_RATE / 100;
-      discountRate = 1 - totalAmount / originalTotal;
-    }
-
-    return {
-      totalAmount: Math.round(totalAmount),
-      discountRate,
-      originalTotal,
-      isSpecialDiscount,
-    };
-  },
-
-  /**
-   * 장바구니 전체 계산 실행 및 상태 업데이트
-   * @param {HTMLCollection} cartItems - 장바구니 DOM 아이템들
-   * @returns {Object} 계산 결과
-   */
-  updateCartCalculation(cartItems) {
-    const basicCalculation = this.calculateCartTotals(cartItems);
-
-    const finalCalculation = this.calculateFinalAmount(
-      basicCalculation.subtotal,
-      basicCalculation.itemCount,
-      basicCalculation.itemDiscounts,
-    );
-
-    this.setCartTotals(finalCalculation.totalAmount, basicCalculation.itemCount);
-
-    return {
-      ...basicCalculation,
-      ...finalCalculation,
-    };
-  },
-};
-
-const useBonusPointsManager = {
-  bonusPoints: 0,
-
-  /**
-   * 현재 보너스 포인트 반환
-   * @returns {number} 보너스 포인트
-   */
-  getBonusPoints() {
-    return this.bonusPoints;
-  },
-
-  /**
-   * 보너스 포인트 설정
-   * @param {number} points - 설정할 포인트
-   */
-  setBonusPoints(points) {
-    this.bonusPoints = points;
-  },
-
-  /**
-   * 보너스 포인트 초기화
-   */
-  resetBonusPoints() {
-    this.bonusPoints = 0;
-  },
-
-  /**
-   * 기본 포인트 계산 (구매액 기준)
-   * @param {number} totalAmount - 총 구매액
-   * @returns {number} 기본 포인트
-   */
-  calculateBasePoints(totalAmount) {
-    return Math.floor(totalAmount / POINTS_RULES.BASE_CALCULATION_UNIT);
-  },
-
-  /**
-   * 특별 포인트 날짜 보너스 계산
-   * @param {number} basePoints - 기본 포인트
-   * @returns {Object} {points, isSpecialDay, detail}
-   */
-  calculateSpecialDayBonus(basePoints) {
-    const isSpecialDay = isSpecialPointsDay();
-
-    if (isSpecialDay && basePoints > 0) {
-      const bonusPoints = basePoints * POINTS_RULES.SPECIAL_POINTS_MULTIPLIER;
-      const detail = `${POINTS_RULES.SPECIAL_POINTS_DAYS.map(getKoreanDayName).join(", ")} ${POINTS_RULES.SPECIAL_POINTS_MULTIPLIER}배`;
-
-      return {
-        points: bonusPoints,
-        isSpecialDay: true,
-        detail,
-      };
-    }
-
-    return {
-      points: basePoints,
-      isSpecialDay: false,
-      detail: basePoints > 0 ? `기본: ${basePoints}p` : "",
-    };
-  },
-
-  /**
-   * 장바구니 아이템들로부터 콤보 보너스 계산
-   * @param {HTMLCollection} cartItems - 장바구니 DOM 아이템들
-   * @returns {Object} {bonusPoints, details}
-   */
-  calculateComboBonus(cartItems: HTMLCollection) {
-    let bonusPoints = 0;
-    const details: string[] = [];
-
-    let hasKeyboard = false;
-    let hasMouse = false;
-    let hasMonitorArm = false;
-
-    Array.from(cartItems).forEach((node) => {
-      const element = node as HTMLElement;
-      const product = useProductData.findProductById(element.id);
-      if (product) {
-        if (product.id === PRODUCT_IDS.KEYBOARD) {
-          hasKeyboard = true;
-        } else if (product.id === PRODUCT_IDS.MOUSE) {
-          hasMouse = true;
-        } else if (product.id === PRODUCT_IDS.MONITOR_ARM) {
-          hasMonitorArm = true;
-        }
-      }
-    });
-
-    if (hasKeyboard && hasMouse) {
-      bonusPoints += POINTS_RULES.COMBO_BONUS.KEYBOARD_MOUSE;
-      details.push(`키보드+마우스 세트 +${POINTS_RULES.COMBO_BONUS.KEYBOARD_MOUSE}p`);
-    }
-
-    if (hasKeyboard && hasMouse && hasMonitorArm) {
-      bonusPoints += POINTS_RULES.COMBO_BONUS.FULL_SET;
-      details.push(`풀세트 구매 +${POINTS_RULES.COMBO_BONUS.FULL_SET}p`);
-    }
-
-    return {
-      bonusPoints,
-      details,
-    };
-  },
-
-  /**
-   * 수량별 보너스 포인트 계산
-   * @param {number} totalItemCount - 총 상품 개수
-   * @returns {Object} {bonusPoints, detail}
-   */
-  calculateQuantityBonus(totalItemCount) {
-    let bonusPoints = 0;
-    let detail = "";
-
-    if (totalItemCount >= POINTS_RULES.QUANTITY_BONUS.TIER_3.threshold) {
-      bonusPoints = POINTS_RULES.QUANTITY_BONUS.TIER_3.bonus;
-      detail = `대량구매(${POINTS_RULES.QUANTITY_BONUS.TIER_3.threshold}개+) +${POINTS_RULES.QUANTITY_BONUS.TIER_3.bonus}p`;
-    } else if (totalItemCount >= POINTS_RULES.QUANTITY_BONUS.TIER_2.threshold) {
-      bonusPoints = POINTS_RULES.QUANTITY_BONUS.TIER_2.bonus;
-      detail = `대량구매(${POINTS_RULES.QUANTITY_BONUS.TIER_2.threshold}개+) +${POINTS_RULES.QUANTITY_BONUS.TIER_2.bonus}p`;
-    } else if (totalItemCount >= POINTS_RULES.QUANTITY_BONUS.TIER_1.threshold) {
-      bonusPoints = POINTS_RULES.QUANTITY_BONUS.TIER_1.bonus;
-      detail = `대량구매(${POINTS_RULES.QUANTITY_BONUS.TIER_1.threshold}개+) +${POINTS_RULES.QUANTITY_BONUS.TIER_1.bonus}p`;
-    }
-
-    return {
-      bonusPoints,
-      detail,
-    };
-  },
-
-  /**
-   * 전체 보너스 포인트 계산 및 업데이트
-   * @param {number} totalAmount - 총 구매액
-   * @param {number} totalItemCount - 총 상품 개수
-   * @param {HTMLCollection} cartItems - 장바구니 DOM 아이템들
-   * @returns {Object} 계산 결과 및 상세 정보
-   */
-  calculateAndUpdateBonusPoints(totalAmount: number, totalItemCount: number, cartItems: HTMLCollection) {
-    const details: string[] = [];
-
-    const basePoints = this.calculateBasePoints(totalAmount);
-
-    const specialDayResult = this.calculateSpecialDayBonus(basePoints);
-    let finalPoints = specialDayResult.points;
-    if (specialDayResult.detail) {
-      details.push(specialDayResult.detail);
-    }
-
-    const comboResult = this.calculateComboBonus(cartItems);
-    finalPoints += comboResult.bonusPoints;
-    details.push(...comboResult.details);
-
-    const quantityResult = this.calculateQuantityBonus(totalItemCount);
-    finalPoints += quantityResult.bonusPoints;
-    if (quantityResult.detail) {
-      details.push(quantityResult.detail);
-    }
-
-    this.setBonusPoints(finalPoints);
-
-    return {
-      totalPoints: finalPoints,
-      details,
-      breakdown: {
-        base: basePoints,
-        specialDay: specialDayResult,
-        combo: comboResult,
-        quantity: quantityResult,
-      },
-    };
-  },
-};
-
-/**
- * 장바구니 아이템들의 할인 표시 스타일 업데이트
- * @param {HTMLCollection} cartItems - 장바구니 DOM 아이템들
- */
-function updateCartItemStyles(cartItems: HTMLCollection): void {
-  for (let i = 0; i < cartItems.length; i += 1) {
-    const cartItem = cartItems[i] as HTMLElement;
-    const curItem = useProductData.findProductById(cartItem.id);
-
-    if (curItem) {
-      const qtyElem = cartItem.querySelector(".quantity-number") as HTMLElement;
-      if (qtyElem?.textContent) {
-        const q = parseInt(qtyElem.textContent, 10);
-        const itemDiv = cartItem;
-        const priceElems = itemDiv.querySelectorAll(".text-lg, .text-xs");
-
-        priceElems.forEach(function (elem) {
-          if (elem.classList.contains("text-lg")) {
-            const newFontWeight = q >= DISCOUNT_RULES.ITEM_DISCOUNT_THRESHOLD ? "bold" : "normal";
-            const targetElement = elem as HTMLElement;
-            if (targetElement.style.fontWeight !== newFontWeight) {
-              targetElement.style.fontWeight = newFontWeight;
-            }
-          }
-        });
-      }
-    }
-  }
-}
-
-/**
- * 화요일 특별 할인 UI 표시 업데이트
- * @param {boolean} isSpecialDiscount - 특별 할인 여부
- * @param {number} totalAmount - 총 금액
- */
-function updateSpecialDiscountDisplay(isSpecialDiscount: boolean, totalAmount: number): void {
-  const tuesdaySpecial = document.getElementById("tuesday-special");
-
-  if (tuesdaySpecial) {
-    if (isSpecialDiscount && totalAmount > 0) {
-      tuesdaySpecial.classList.remove("hidden");
-    } else {
-      tuesdaySpecial.classList.add("hidden");
-    }
-  }
-}
-
-/**
- * 상품 개수 표시 업데이트
- * @param {number} itemCount - 총 상품 개수
- */
-function updateItemCountDisplay(itemCount: number): void {
-  const itemCountElement = document.getElementById("item-count");
-
-  if (itemCountElement && itemCountElement.textContent) {
-    const match = itemCountElement.textContent.match(/\d+/);
-    const previousCount = parseInt(match ? match[0] : "0", 10);
-    itemCountElement.textContent = `🛍️ ${itemCount} items in cart`;
-
-    if (previousCount !== itemCount) {
-      itemCountElement.setAttribute("data-changed", "true");
-    }
-  }
-}
-
-/**
- * 주문 요약 데이터 계산 (순수 함수)
- * @param {HTMLCollection} cartItems - 장바구니 DOM 아이템들
- * @param {number} subtotal - 소계
- * @param {number} itemCount - 총 상품 개수
- * @param {Array} itemDiscounts - 개별 상품 할인 목록
- * @param {boolean} isSpecialDiscount - 특별 할인 여부
- * @param {number} totalAmount - 총 금액
- * @returns {Object} 주문 요약 데이터
- */
-function calculateOrderSummaryData(
-  cartItems: HTMLCollection,
-  subtotal: number,
-  itemCount: number,
-  itemDiscounts: IItemDiscount[],
-  isSpecialDiscount: boolean,
-  totalAmount: number,
-) {
-  const items: ICartItem[] = [];
-  for (let i = 0; i < cartItems.length; i += 1) {
-    const cartItem = cartItems[i] as HTMLElement;
-    const curItem = useProductData.findProductById(cartItem.id);
-    if (curItem) {
-      const qtyElem = cartItem.querySelector(".quantity-number") as HTMLElement;
-      if (qtyElem?.textContent) {
-        const quantity = parseInt(qtyElem.textContent, 10);
-        const itemTotal = curItem.val * quantity;
-
-        items.push({
-          name: curItem.name,
-          quantity,
-          itemTotal,
-        });
-      }
-    }
-  }
-
-  const discounts = {
-    hasBulkDiscount: itemCount >= DISCOUNT_RULES.BULK_DISCOUNT_THRESHOLD,
-    bulkDiscountRate: DISCOUNT_RULES.BULK_DISCOUNT_RATE,
-    bulkDiscountThreshold: DISCOUNT_RULES.BULK_DISCOUNT_THRESHOLD,
-    itemDiscounts,
-    itemDiscountThreshold: DISCOUNT_RULES.ITEM_DISCOUNT_THRESHOLD,
-    hasSpecialDiscount: isSpecialDiscount && totalAmount > 0,
-    specialDiscountDays: DISCOUNT_RULES.SPECIAL_DISCOUNT_DAYS.map(getKoreanDayName).join(", "),
-    specialDiscountRate: DISCOUNT_RULES.SPECIAL_DISCOUNT_RATE,
-  };
-
-  return {
-    items,
-    subtotal,
-    discounts,
-    shouldRender: subtotal > 0,
-  };
-}
-
-/**
- * 주문 요약 HTML 템플릿 생성 (순수 함수)
- * @param {Object} summaryData - 주문 요약 데이터
- * @returns {string} HTML 템플릿 문자열
- */
-function createOrderSummaryHTML(summaryData) {
-  if (!summaryData.shouldRender) {
-    return "";
-  }
-
-  const itemsHTML = summaryData.items
-    .map(
-      (item) => `
-      <div class="flex justify-between text-xs tracking-wide text-gray-400">
-        <span>${item.name} x ${item.quantity}</span>
-        <span>₩${item.itemTotal.toLocaleString()}</span>
-      </div>
-    `,
-    )
-    .join("");
-
-  const subtotalHTML = `
-    <div class="border-t border-white/10 my-3"></div>
-    <div class="flex justify-between text-sm tracking-wide">
-      <span>Subtotal</span>
-      <span>₩${summaryData.subtotal.toLocaleString()}</span>
-    </div>
-  `;
-
-  let discountsHTML = "";
-
-  if (summaryData.discounts.hasBulkDiscount) {
-    discountsHTML += `
-      <div class="flex justify-between text-sm tracking-wide text-green-400">
-        <span class="text-xs">🎉 대량구매 할인 (${summaryData.discounts.bulkDiscountThreshold}개 이상)</span>
-        <span class="text-xs">-${summaryData.discounts.bulkDiscountRate}%</span>
-      </div>
-    `;
-  } else if (summaryData.discounts.itemDiscounts.length > 0) {
-    summaryData.discounts.itemDiscounts.forEach(function (item) {
-      discountsHTML += `
-        <div class="flex justify-between text-sm tracking-wide text-green-400">
-          <span class="text-xs">${item.name} (${summaryData.discounts.itemDiscountThreshold}개↑)</span>
-          <span class="text-xs">-${item.discount}%</span>
-        </div>
-      `;
-    });
-  }
-
-  let specialDiscountHTML = "";
-  if (summaryData.discounts.hasSpecialDiscount) {
-    specialDiscountHTML = `
-      <div class="flex justify-between text-sm tracking-wide text-purple-400">
-        <span class="text-xs">🌟 ${summaryData.discounts.specialDiscountDays} 추가 할인</span>
-        <span class="text-xs">-${summaryData.discounts.specialDiscountRate}%</span>
-      </div>
-    `;
-  }
-
-  return itemsHTML + subtotalHTML + discountsHTML + specialDiscountHTML;
-}
-
-/**
- * 주문 요약 렌더러 객체
- * DOM 조작만 담당
- */
-const OrderSummaryRenderer = {
-  /**
-   * 주문 요약 렌더링
-   * @param {Object} summaryData - 주문 요약 데이터
-   */
-  render(summaryData: any): void {
-    const summaryDetails = document.getElementById("summary-details");
-    if (!summaryDetails) return;
-
-    summaryDetails.innerHTML = "";
-
-    if (!summaryData.shouldRender) {
-      return;
-    }
-
-    const html = createOrderSummaryHTML(summaryData);
-    summaryDetails.innerHTML = html;
-  },
-};
-
-/**
- * 주문 요약 세부 정보 렌더링 (리팩토링된 버전)
- * @param {HTMLCollection} cartItems - 장바구니 DOM 아이템들
- * @param {number} subtotal - 소계
- * @param {number} itemCount - 총 상품 개수
- * @param {Array} itemDiscounts - 개별 상품 할인 목록
- * @param {boolean} isSpecialDiscount - 특별 할인 여부
- * @param {number} totalAmount - 총 금액
- */
-function renderOrderSummaryDetails(cartItems, subtotal, itemCount, itemDiscounts, isSpecialDiscount, totalAmount) {
-  const summaryData = calculateOrderSummaryData(
-    cartItems,
-    subtotal,
-    itemCount,
-    itemDiscounts,
-    isSpecialDiscount,
-    totalAmount,
-  );
-  OrderSummaryRenderer.render(summaryData);
-}
-
-/**
- * 총액 및 포인트 표시 데이터 계산 (순수 함수)
- * @param {number} totalAmount - 총 금액
- * @returns {Object} 총액 및 포인트 표시 데이터
- */
-function calculateTotalAndPointsData(totalAmount) {
-  const formattedTotal = `₩${totalAmount.toLocaleString()}`;
-  const points = Math.floor(totalAmount / POINTS_RULES.BASE_CALCULATION_UNIT);
-
-  return {
-    totalText: formattedTotal,
-    points,
-    pointsText: `적립 포인트: ${points}p`,
-    shouldShowPoints: true,
-  };
-}
-
-/**
- * 총액 및 포인트 렌더러 객체
- * DOM 조작만 담당
- */
-const TotalPointsRenderer = {
-  /**
-   * 총액 및 포인트 표시 렌더링
-   * @param {Object} displayData - 표시 데이터
-   */
-  render(displayData) {
-    const cartSummaryElement = useDOMManager.getElement("cartSummary");
-
-    const totalDiv = cartSummaryElement.querySelector(".text-2xl");
-    const loyaltyPointsDiv = document.getElementById("loyalty-points");
-
-    if (totalDiv) {
-      totalDiv.textContent = displayData.totalText;
-    }
-
-    if (loyaltyPointsDiv && displayData.shouldShowPoints) {
-      loyaltyPointsDiv.textContent = displayData.pointsText;
-      loyaltyPointsDiv.style.display = "block";
-    }
-  },
-};
-
-/**
- * 총액 및 포인트 표시 업데이트 (리팩토링된 버전)
- * @param {number} totalAmount - 총 금액
- */
-function updateTotalAndPointsDisplay(totalAmount) {
-  const displayData = calculateTotalAndPointsData(totalAmount);
-  TotalPointsRenderer.render(displayData);
-}
-
-/**
- * 할인 정보 데이터 계산 (순수 함수)
- * @param {number} discountRate - 할인율
- * @param {number} totalAmount - 총 금액
- * @param {number} originalTotal - 원래 총액
- * @returns {Object} 할인 정보 데이터
- */
-function calculateDiscountInfoData(discountRate, totalAmount, originalTotal) {
-  const hasDiscount = discountRate > 0 && totalAmount > 0;
-  const savedAmount = hasDiscount ? originalTotal - totalAmount : 0;
-  const discountPercentage = hasDiscount ? (discountRate * 100).toFixed(1) : "0.0";
-
-  return {
-    hasDiscount,
-    savedAmount,
-    discountPercentage,
-    formattedSavedAmount: `₩${Math.round(savedAmount).toLocaleString()}`,
-  };
-}
-
-/**
- * 할인 정보 HTML 템플릿 생성 (순수 함수)
- * @param {Object} discountData - 할인 정보 데이터
- * @returns {string} HTML 템플릿 문자열
- */
-function createDiscountInfoHTML(discountData) {
-  if (!discountData.hasDiscount) {
-    return "";
-  }
-
-  return `
-    <div class="bg-green-500/20 rounded-lg p-3">
-      <div class="flex justify-between items-center mb-1">
-        <span class="text-xs uppercase tracking-wide text-green-400">총 할인율</span>
-        <span class="text-sm font-medium text-green-400">${discountData.discountPercentage}%</span>
-      </div>
-      <div class="text-2xs text-gray-300">${discountData.formattedSavedAmount} 할인되었습니다</div>
-    </div>
-  `;
-}
-
-/**
- * 할인 정보 렌더러 객체
- * DOM 조작만 담당
- */
-const DiscountInfoRenderer = {
-  /**
-   * 할인 정보 패널 렌더링
-   * @param {Object} discountData - 할인 정보 데이터
-   */
-  render(discountData: IDiscountData): void {
-    const discountInfoDiv = document.getElementById("discount-info");
-    if (!discountInfoDiv) return;
-
-    discountInfoDiv.innerHTML = "";
-
-    if (discountData.hasDiscount) {
-      const html = createDiscountInfoHTML(discountData);
-      discountInfoDiv.innerHTML = html;
-    }
-  },
-};
-
-/**
- * 할인 정보 패널 렌더링 (리팩토링된 버전)
- * @param {number} discountRate - 할인율
- * @param {number} totalAmount - 총 금액
- * @param {number} originalTotal - 원래 총액
- */
-function renderDiscountInfoPanel(discountRate, totalAmount, originalTotal) {
-  const discountData = calculateDiscountInfoData(discountRate, totalAmount, originalTotal);
-  DiscountInfoRenderer.render(discountData);
-}
-
-/**
- * 상품 선택 옵션 데이터 계산 (순수 함수)
- * @returns {Object} 상품 선택 옵션 데이터
- */
-function calculateProductSelectData() {
-  const products = useProductData.getProducts();
-
-  let totalStock = 0;
-  for (let idx = 0; idx < products.length; idx += 1) {
-    const product = products[idx];
-    totalStock += product.q;
-  }
-
-  const optionData = products.map(function (item) {
-    let discountText = "";
-    if (item.onSale) discountText += " ⚡SALE";
-    if (item.suggestSale) discountText += " 💝추천";
-
-    let optionText;
-    let optionClass;
-    let isDisabled;
-
-    if (item.q === 0) {
-      optionText = `${item.name} - ${item.val}원 (품절)${discountText}`;
-      optionClass = "text-gray-400";
-      isDisabled = true;
-    } else if (item.onSale && item.suggestSale) {
-      const totalDiscountRate = DISCOUNT_RULES.LIGHTNING_SALE_RATE + DISCOUNT_RULES.RECOMMENDATION_DISCOUNT_RATE;
-      optionText = `⚡💝${item.name} - ${item.originalVal}원 → ${item.val}원 (${totalDiscountRate}% SUPER SALE!)`;
-      optionClass = "text-purple-600 font-bold";
-      isDisabled = false;
-    } else if (item.onSale) {
-      optionText = `⚡${item.name} - ${item.originalVal}원 → ${item.val}원 (${DISCOUNT_RULES.LIGHTNING_SALE_RATE}% SALE!)`;
-      optionClass = "text-red-500 font-bold";
-      isDisabled = false;
-    } else if (item.suggestSale) {
-      optionText = `💝${item.name} - ${item.originalVal}원 → ${item.val}원 (${DISCOUNT_RULES.RECOMMENDATION_DISCOUNT_RATE}% 추천할인!)`;
-      optionClass = "text-blue-500 font-bold";
-      isDisabled = false;
-    } else {
-      optionText = `${item.name} - ${item.val}원${discountText}`;
-      optionClass = "";
-      isDisabled = false;
-    }
-
-    return {
-      id: item.id,
-      text: optionText,
-      className: optionClass,
-      disabled: isDisabled,
-    };
-  });
-
-  return {
-    options: optionData,
-    totalStock,
-    shouldShowWarning: totalStock < STOCK_THRESHOLDS.TOTAL_STOCK_WARNING,
-  };
-}
-
-/**
- * 상품 선택 옵션 엘리먼트들 생성 (순수 함수)
- * @param {Object} selectData - 상품 선택 데이터
- * @returns {Array} option 엘리먼트 배열
- */
-function createProductSelectOptions(selectData) {
-  return selectData.options.map(function (optionData) {
-    const opt = document.createElement("option");
-    opt.value = optionData.id;
-    opt.textContent = optionData.text;
-    opt.disabled = optionData.disabled;
-    if (optionData.className) {
-      opt.className = optionData.className;
-    }
-    return opt;
-  });
-}
-
-/**
- * 상품 선택 렌더러 객체
- * DOM 조작만 담당
- */
-const ProductSelectRenderer = {
-  /**
-   * 상품 선택 옵션 렌더링
-   * @param {Object} selectData - 상품 선택 데이터
-   */
-  render(selectData) {
-    const productSelectElement = useDOMManager.getElement("productSelect");
-
-    productSelectElement.innerHTML = "";
-
-    const options = createProductSelectOptions(selectData);
-    options.forEach(function (opt) {
-      productSelectElement.appendChild(opt);
-    });
-
-    if (selectData.shouldShowWarning) {
-      productSelectElement.style.borderColor = "orange";
-    } else {
-      productSelectElement.style.borderColor = "";
-    }
-  },
-};
-
-/**
- * 상품 선택 옵션 업데이트 (리팩토링된 버전)
- */
-function updateProductSelectOptions() {
-  const selectData = calculateProductSelectData();
-  ProductSelectRenderer.render(selectData);
-}
-
-function renderBonusPointsDisplay(): void {
-  const cartDisplayElement = useDOMManager.getElement("cartDisplay");
-  if (!cartDisplayElement) return;
-
-  const totalAmount = useCartManager.getTotalAmount();
-  const itemCount = useCartManager.getItemCount();
-  const nodes = cartDisplayElement.children;
-
-  if (cartDisplayElement.children.length === 0) {
-    const loyaltyPoints = document.getElementById("loyalty-points");
-    if (loyaltyPoints) {
-      loyaltyPoints.style.display = "none";
-    }
-    useBonusPointsManager.resetBonusPoints();
-    return;
-  }
-
-  const bonusResult = useBonusPointsManager.calculateAndUpdateBonusPoints(totalAmount, itemCount, nodes);
-
-  const finalPoints = bonusResult.totalPoints;
-  const pointsDetail = bonusResult.details;
-
-  const ptsTag = document.getElementById("loyalty-points");
-  if (ptsTag) {
-    if (finalPoints > 0) {
-      ptsTag.innerHTML =
-        `<div>적립 포인트: <span class="font-bold">${finalPoints}p</span></div>` +
-        `<div class="text-2xs opacity-70 mt-1">${pointsDetail.join(", ")}</div>`;
-      ptsTag.style.display = "block";
-    } else {
-      ptsTag.textContent = "적립 포인트: 0p";
-      ptsTag.style.display = "block";
-    }
-  }
-}
-
-/**
- * 장바구니 계산 및 전체 UI 업데이트
- * 메인 함수: 장바구니 상태 계산 후 모든 UI 컴포넌트 업데이트
- */
-function updateCartDisplay() {
-  const cartDisplayElement = useDOMManager.getElement("cartDisplay");
-  const cartItems = cartDisplayElement.children;
-
-  const calculation = useCartManager.updateCartCalculation(cartItems);
-  const { subtotal, itemCount, totalAmount, discountRate, originalTotal, isSpecialDiscount, itemDiscounts } =
-    calculation;
-
-  updateCartItemStyles(cartItems);
-  updateSpecialDiscountDisplay(isSpecialDiscount, totalAmount);
-  updateItemCountDisplay(itemCount);
-  renderOrderSummaryDetails(cartItems, subtotal, itemCount, itemDiscounts, isSpecialDiscount, totalAmount);
-  updateTotalAndPointsDisplay(totalAmount);
-  renderDiscountInfoPanel(discountRate, totalAmount, originalTotal);
-
-  useStockManager.updateStockInfoDisplay();
-  renderBonusPointsDisplay();
-}
-
-/**
- * 장바구니 아이템 가격 표시 데이터 계산 (순수 함수)
- * @param {HTMLCollection} cartItems - 장바구니 DOM 아이템들
- * @returns {Array} 아이템별 가격 표시 데이터 배열
- */
-function calculateCartItemPricesData(cartItems) {
-  const itemsData = [];
-
-  for (let i = 0; i < cartItems.length; i += 1) {
-    const itemId = cartItems[i].id;
-    const product = useProductData.findProductById(itemId);
-
-    if (product) {
-      let priceHTML;
-      let nameText;
-      let priceClassName;
-
-      if (product.onSale && product.suggestSale) {
-        priceHTML = `<span class="line-through text-gray-400">₩${product.originalVal.toLocaleString()}</span> <span class="text-purple-600">₩${product.val.toLocaleString()}</span>`;
-        nameText = `⚡💝${product.name}`;
-        priceClassName = "text-purple-600";
-      } else if (product.onSale) {
-        priceHTML = `<span class="line-through text-gray-400">₩${product.originalVal.toLocaleString()}</span> <span class="text-red-500">₩${product.val.toLocaleString()}</span>`;
-        nameText = `⚡${product.name}`;
-        priceClassName = "text-red-500";
-      } else if (product.suggestSale) {
-        priceHTML = `<span class="line-through text-gray-400">₩${product.originalVal.toLocaleString()}</span> <span class="text-blue-500">₩${product.val.toLocaleString()}</span>`;
-        nameText = `💝${product.name}`;
-        priceClassName = "text-blue-500";
-      } else {
-        priceHTML = `₩${product.val.toLocaleString()}`;
-        nameText = product.name;
-        priceClassName = "";
-      }
-
-      itemsData.push({
-        itemIndex: i,
-        priceHTML,
-        nameText,
-        priceClassName,
-        isDiscounted: product.onSale || product.suggestSale,
-      });
-    }
-  }
-
-  return itemsData;
-}
-
-/**
- * 장바구니 아이템 가격 렌더러 객체
- * DOM 조작만 담당
- */
-const CartItemPricesRenderer = {
-  /**
-   * 장바구니 아이템 가격 표시 렌더링
-   * @param {HTMLCollection} cartItems - 장바구니 DOM 아이템들
-   * @param {Array} itemsData - 아이템별 가격 표시 데이터
-   */
-  render(cartItems: HTMLCollection, itemsData: ICartItemData[]): void {
-    itemsData.forEach(function (itemData) {
-      const cartItem = cartItems[itemData.itemIndex];
-      const priceDiv = cartItem.querySelector(".text-lg");
-      const nameDiv = cartItem.querySelector("h3");
-
-      if (priceDiv) {
-        if (itemData.isDiscounted) {
-          priceDiv.innerHTML = itemData.priceHTML;
-        } else {
-          priceDiv.textContent = itemData.priceHTML;
-        }
-      }
-
-      if (nameDiv) {
-        nameDiv.textContent = itemData.nameText;
-      }
-    });
-  },
-};
-
-/**
- * 장바구니 아이템 가격 업데이트 (리팩토링된 버전)
- */
-function updateCartItemPrices() {
-  const cartDisplayElement = useDOMManager.getElement("cartDisplay");
-  const cartItems = cartDisplayElement.children;
-  const itemsData = calculateCartItemPricesData(cartItems);
-
-  CartItemPricesRenderer.render(cartItems, itemsData);
-  updateCartDisplay();
-}
+import { IProduct } from "./types";
+import { SALE_INTERVALS, DISCOUNT_RULES } from "./constants";
+import { useDOMManager } from "./utils/domManager";
+import {
+  validateAddToCartInput,
+  calculateItemDisplayData,
+  createCartItemHTML,
+  parseCartClickEvent,
+  calculateQuantityChange,
+} from "./utils/eventHandlers";
+import { updateCartDisplay, updateProductSelectOptions, updateCartItemPrices } from "./utils/uiUpdaters";
+
+import { useProductData } from "./domains/products/productData";
+import { useCartManager } from "./domains/cart/cartManager";
 
 /**
  * 앱 상태 초기화
@@ -1341,6 +28,7 @@ function initializeAppState() {
  */
 function createMainLayoutElements() {
   const root = document.getElementById("app");
+  if (!root) throw new Error("App root element not found");
 
   const header = document.createElement("div");
   header.className = "mb-8";
@@ -1434,7 +122,7 @@ function createRightColumnElements() {
     </p>
   `;
 
-  const cartSummaryElement = rightColumn.querySelector("#cart-total");
+  const cartSummaryElement = rightColumn.querySelector("#cart-total") as HTMLElement;
   useDOMManager.setElement("cartSummary", cartSummaryElement);
 
   return rightColumn;
@@ -1553,8 +241,8 @@ function startSpecialSaleTimers() {
       if (luckyItem && luckyItem.q > 0 && !luckyItem.onSale) {
         const discountedPrice = Math.round((luckyItem.originalVal * (100 - DISCOUNT_RULES.LIGHTNING_SALE_RATE)) / 100);
         useProductData.updateProductSaleStatus(luckyItem.id, {
-          val: discountedPrice,
           onSale: true,
+          suggestSale: false,
         });
 
         alert(`⚡번개세일! ${luckyItem.name}이(가) ${DISCOUNT_RULES.LIGHTNING_SALE_RATE}% 할인 중입니다!`);
@@ -1574,9 +262,9 @@ function startSpecialSaleTimers() {
       }
 
       const products = useProductData.getProducts();
-      const isNotLastSelected = (product) => product.id !== lastSelectedProductId;
-      const isInStock = (product) => product.q > 0;
-      const isNotSuggested = (product) => !product.suggestSale;
+      const isNotLastSelected = (product: IProduct) => product.id !== lastSelectedProductId;
+      const isInStock = (product: IProduct) => product.q > 0;
+      const isNotSuggested = (product: IProduct) => !product.suggestSale;
 
       const suggest = products.find(
         (product) => isNotLastSelected(product) && isInStock(product) && isNotSuggested(product),
@@ -1619,252 +307,99 @@ function main() {
   startSpecialSaleTimers();
 }
 
-/**
- * 장바구니 추가 입력 검증 (순수 함수)
- * @param {string} selectedId - 선택된 상품 ID
- * @param {Object} product - 상품 객체
- * @returns {Object} 검증 결과
- */
-function validateAddToCartInput(selectedId: string, product: IProduct | null): { isValid: boolean; reason?: string } {
-  if (!selectedId || !product) {
-    return { isValid: false, reason: "INVALID_SELECTION" };
-  }
-  if (product.q <= 0) {
-    return { isValid: false, reason: "OUT_OF_STOCK" };
-  }
-  return { isValid: true };
-}
+// 이벤트 리스너 설정
+function setupEventListeners() {
+  const addToCartButton = useDOMManager.getElement("addToCartButton") as HTMLButtonElement;
+  const productSelectElement = useDOMManager.getElement("productSelect") as HTMLSelectElement;
+  const cartDisplayElement = useDOMManager.getElement("cartDisplay") as HTMLDivElement;
 
-/**
- * 상품 표시 데이터 계산 (순수 함수)
- * @param {Object} product - 상품 객체
- * @returns {Object} 표시 데이터
- */
-function calculateItemDisplayData(product: IProduct): {
-  titlePrefix: string;
-  priceDisplay: string;
-  priceClass: string;
-  name: string;
-  id: string;
-  val: number;
-  originalVal: number;
-} {
-  let titlePrefix = "";
-  if (product.onSale && product.suggestSale) {
-    titlePrefix = "⚡💝";
-  } else if (product.onSale) {
-    titlePrefix = "⚡";
-  } else if (product.suggestSale) {
-    titlePrefix = "💝";
-  }
-
-  let priceDisplay = "";
-  let priceClass = "";
-  if (product.onSale || product.suggestSale) {
-    if (product.onSale && product.suggestSale) {
-      priceClass = "text-purple-600";
-    } else if (product.onSale) {
-      priceClass = "text-red-500";
-    } else {
-      priceClass = "text-blue-500";
-    }
-    priceDisplay = `<span class="line-through text-gray-400">₩${product.originalVal.toLocaleString()}</span> <span class="${priceClass}">₩${product.val.toLocaleString()}</span>`;
-  } else {
-    priceDisplay = `₩${product.val.toLocaleString()}`;
-  }
-
-  return {
-    titlePrefix,
-    priceDisplay,
-    priceClass,
-    name: product.name,
-    id: product.id,
-    val: product.val,
-    originalVal: product.originalVal,
-  };
-}
-
-/**
- * 장바구니 아이템 HTML 템플릿 생성 (순수 함수)
- * @param {Object} itemData - 아이템 표시 데이터
- * @returns {string} HTML 템플릿
- */
-function createCartItemHTML(itemData: { titlePrefix: string; name: string; priceDisplay: string; id: string }): string {
-  return `
-    <div class="w-20 h-20 bg-gradient-black relative overflow-hidden">
-      <div class="absolute top-1/2 left-1/2 w-[60%] h-[60%] bg-white/10 -translate-x-1/2 -translate-y-1/2 rotate-45"></div>
-    </div>
-    <div>
-      <h3 class="text-base font-normal mb-1 tracking-tight">${itemData.titlePrefix}${itemData.name}</h3>
-      <p class="text-xs text-gray-500 mb-0.5 tracking-wide">PRODUCT</p>
-      <p class="text-xs text-black mb-3">${itemData.priceDisplay}</p>
-      <div class="flex items-center gap-4">
-        <button class="quantity-change w-6 h-6 border border-black bg-white text-sm flex items-center justify-center transition-all hover:bg-black hover:text-white" data-product-id="${itemData.id}" data-change="-1">−</button>
-        <span class="quantity-number text-sm font-normal min-w-[20px] text-center tabular-nums">1</span>
-        <button class="quantity-change w-6 h-6 border border-black bg-white text-sm flex items-center justify-center transition-all hover:bg-black hover:text-white" data-product-id="${itemData.id}" data-change="1">+</button>
-      </div>
-    </div>
-    <div class="text-right">
-      <div class="text-lg mb-2 tracking-tight tabular-nums">${itemData.priceDisplay}</div>
-      <a class="remove-item text-2xs text-gray-500 uppercase tracking-wider cursor-pointer transition-colors border-b border-transparent hover:text-black hover:border-black" data-product-id="${itemData.id}">Remove</a>
-    </div>
-  `;
-}
-
-/**
- * 장바구니 이벤트 타입 결정 (순수 함수)
- * @param {Event} event - 클릭 이벤트
- * @returns {Object} 이벤트 처리 정보
- */
-function parseCartClickEvent(event: MouseEvent): {
-  shouldHandle: boolean;
-  actionType?: string;
-  productId?: string;
-  quantityChange?: number;
-} {
-  const target = event.target as HTMLElement;
-
-  if (!target.classList.contains("quantity-change") && !target.classList.contains("remove-item")) {
-    return { shouldHandle: false };
-  }
-
-  const { productId } = target.dataset;
-  if (!productId) {
-    return { shouldHandle: false };
-  }
-
-  if (target.classList.contains("quantity-change")) {
-    return {
-      shouldHandle: true,
-      actionType: "QUANTITY_CHANGE",
-      productId,
-      quantityChange: parseInt(target.dataset.change, 10),
-    };
-  }
-
-  if (target.classList.contains("remove-item")) {
-    return {
-      shouldHandle: true,
-      actionType: "REMOVE_ITEM",
-      productId,
-    };
-  }
-
-  return { shouldHandle: false };
-}
-
-/**
- * 수량 변경 계산 (순수 함수)
- * @param {number} currentQuantity - 현재 수량
- * @param {number} quantityChange - 수량 변경값
- * @param {number} availableStock - 사용 가능한 재고
- * @returns {Object} 수량 변경 결과
- */
-function calculateQuantityChange(currentQuantity, quantityChange, availableStock) {
-  const newQuantity = currentQuantity + quantityChange;
-
-  if (newQuantity > 0 && newQuantity <= availableStock + currentQuantity) {
-    return {
-      isValid: true,
-      action: "UPDATE_QUANTITY",
-      newQuantity,
-      stockChange: -quantityChange,
-    };
-  }
-
-  if (newQuantity <= 0) {
-    return {
-      isValid: true,
-      action: "REMOVE_ITEM",
-      stockChange: currentQuantity,
-    };
-  }
-
-  return {
-    isValid: false,
-    reason: "INSUFFICIENT_STOCK",
-    message: "재고가 부족합니다.",
-  };
-}
-
-main();
-
-const addToCartButton = useDOMManager.getElement("addToCartButton");
-const productSelectElement = useDOMManager.getElement("productSelect");
-
-addToCartButton.addEventListener("click", function () {
-  const selItem = productSelectElement.value;
-  const itemToAdd = useProductData.findProductById(selItem);
-
-  const validation = validateAddToCartInput(selItem, itemToAdd);
-  if (!validation.isValid) {
+  if (!addToCartButton || !productSelectElement || !cartDisplayElement) {
+    console.error("Required DOM elements not found");
     return;
   }
 
-  if (itemToAdd && itemToAdd.q > 0) {
-    const item = document.getElementById(itemToAdd.id);
-    if (item) {
-      const qtyElem = item.querySelector(".quantity-number");
-      const newQty = parseInt(qtyElem.textContent, 10) + 1;
-      if (newQty <= itemToAdd.q + parseInt(qtyElem.textContent, 10)) {
-        qtyElem.textContent = newQty;
-        useProductData.updateProductStock(itemToAdd.id, -1);
+  addToCartButton.addEventListener("click", function () {
+    const selItem = productSelectElement.value;
+    const itemToAdd = useProductData.findProductById(selItem);
+
+    const validation = validateAddToCartInput(selItem, itemToAdd);
+    if (!validation.isValid) {
+      return;
+    }
+
+    if (itemToAdd && itemToAdd.q > 0) {
+      const item = document.getElementById(itemToAdd.id);
+      if (item) {
+        const qtyElem = item.querySelector(".quantity-number") as HTMLElement;
+        if (qtyElem?.textContent) {
+          const newQty = parseInt(qtyElem.textContent, 10) + 1;
+          if (newQty <= itemToAdd.q + parseInt(qtyElem.textContent, 10)) {
+            qtyElem.textContent = newQty.toString();
+            useProductData.updateProductStock(itemToAdd.id, -1);
+          } else {
+            alert("재고가 부족합니다.");
+          }
+        }
       } else {
-        alert("재고가 부족합니다.");
+        const newItem = document.createElement("div");
+        newItem.id = itemToAdd.id;
+        newItem.className =
+          "grid grid-cols-[80px_1fr_auto] gap-5 py-5 border-b border-gray-100 first:pt-0 last:border-b-0 last:pb-0";
+
+        const itemDisplayData = calculateItemDisplayData(itemToAdd);
+
+        newItem.innerHTML = createCartItemHTML(itemDisplayData);
+
+        cartDisplayElement.appendChild(newItem);
+        useProductData.updateProductStock(itemToAdd.id, -1);
       }
-    } else {
-      const newItem = document.createElement("div");
-      newItem.id = itemToAdd.id;
-      newItem.className =
-        "grid grid-cols-[80px_1fr_auto] gap-5 py-5 border-b border-gray-100 first:pt-0 last:border-b-0 last:pb-0";
+      updateCartDisplay();
 
-      const itemDisplayData = calculateItemDisplayData(itemToAdd);
-
-      newItem.innerHTML = createCartItemHTML(itemDisplayData);
-
-      const cartDisplayElement = useDOMManager.getElement("cartDisplay");
-      cartDisplayElement.appendChild(newItem);
-      useProductData.updateProductStock(itemToAdd.id, -1);
+      useDOMManager.setState("lastSelectedProductId", selItem);
     }
+  });
+
+  cartDisplayElement.addEventListener("click", function (event) {
+    const eventInfo = parseCartClickEvent(event);
+    if (!eventInfo.shouldHandle) return;
+
+    const itemElem = document.getElementById(eventInfo.productId!);
+    const prod = useProductData.findProductById(eventInfo.productId!);
+    if (!prod || !itemElem) return;
+
+    if (eventInfo.actionType === "QUANTITY_CHANGE") {
+      const qtyElem = itemElem.querySelector(".quantity-number") as HTMLElement;
+      if (!qtyElem?.textContent) return;
+
+      const currentQty = parseInt(qtyElem.textContent, 10);
+
+      const changeResult = calculateQuantityChange(currentQty, eventInfo.quantityChange!, prod.q);
+
+      if (changeResult.isValid) {
+        if (changeResult.action === "UPDATE_QUANTITY") {
+          qtyElem.textContent = changeResult.newQuantity!.toString();
+          useProductData.updateProductStock(prod.id, changeResult.stockChange);
+        } else if (changeResult.action === "REMOVE_ITEM") {
+          useProductData.updateProductStock(prod.id, changeResult.stockChange);
+          itemElem.remove();
+        }
+      } else {
+        alert(changeResult.message);
+      }
+    } else if (eventInfo.actionType === "REMOVE_ITEM") {
+      const qtyElem = itemElem.querySelector(".quantity-number") as HTMLElement;
+      if (!qtyElem?.textContent) return;
+
+      const remQty = parseInt(qtyElem.textContent, 10);
+      useProductData.updateProductStock(prod.id, remQty);
+      itemElem.remove();
+    }
+
     updateCartDisplay();
+    updateProductSelectOptions();
+  });
+}
 
-    useDOMManager.setState("lastSelectedProductId", selItem);
-  }
-});
-
-const cartDisplayElement = useDOMManager.getElement("cartDisplay");
-cartDisplayElement.addEventListener("click", function (event) {
-  const eventInfo = parseCartClickEvent(event);
-  if (!eventInfo.shouldHandle) return;
-
-  const itemElem = document.getElementById(eventInfo.productId);
-  const prod = useProductData.findProductById(eventInfo.productId);
-  if (!prod || !itemElem) return;
-
-  if (eventInfo.actionType === "QUANTITY_CHANGE") {
-    const qtyElem = itemElem.querySelector(".quantity-number");
-    const currentQty = parseInt(qtyElem.textContent, 10);
-
-    const changeResult = calculateQuantityChange(currentQty, eventInfo.quantityChange, prod.q);
-
-    if (changeResult.isValid) {
-      if (changeResult.action === "UPDATE_QUANTITY") {
-        qtyElem.textContent = changeResult.newQuantity;
-        useProductData.updateProductStock(prod.id, changeResult.stockChange);
-      } else if (changeResult.action === "REMOVE_ITEM") {
-        useProductData.updateProductStock(prod.id, changeResult.stockChange);
-        itemElem.remove();
-      }
-    } else {
-      alert(changeResult.message);
-    }
-  } else if (eventInfo.actionType === "REMOVE_ITEM") {
-    const qtyElem = itemElem.querySelector(".quantity-number");
-    const remQty = parseInt(qtyElem.textContent, 10);
-    useProductData.updateProductStock(prod.id, remQty);
-    itemElem.remove();
-  }
-
-  updateCartDisplay();
-  updateProductSelectOptions();
-});
+// 앱 시작
+main();
+setupEventListeners();
